controller 작업

+++ src/main/java/com/klk/bank/controller/ProductController.java
@@ -1,5 +1,6 @@
 package com.klk.bank.controller;
 
+import java.security.Principal;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +9,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.klk.bank.domain.ProductDto;
@@ -27,8 +29,8 @@ public class ProductController {
 	
 	// 상품 조회
 	@GetMapping("search")
-	public void searchPage(String keyword, String type, Model model) {
-		List<ProductDto> list = productService.listProduct(type, keyword);
+	public void searchPage(@RequestParam(name = "keyword", defaultValue = "") String keyword, Model model) {
+		List<ProductDto> list = productService.listProduct(keyword);
 		
 		model.addAttribute("product_list", list);
 	}	
@@ -40,7 +42,9 @@ public class ProductController {
 	}
 	
 	@PostMapping("registration")
-	public String registrationPage(ProductDto product, RedirectAttributes rttr) {
+	public String registrationPage(ProductDto product, Principal principal, RedirectAttributes rttr) {
+				
+		product.setUser_id(principal.getName());
 		boolean success = productService.insertProduct(product);
 		
 		if (success) {
@@ -59,7 +63,7 @@ public class ProductController {
 		ProductDto product = productService.getProductById(id);
 		List<ReplyDto> replyList = replyService.getReplyByProductId(id);
 		model.addAttribute("product", product);
-		model.addAttribute("replyList", replyList);
+		model.addAttribute("reply_list", replyList);
 		
 	}
 
@@ -70,15 +74,27 @@ public class ProductController {
 	}
 	
 	@PostMapping("edit")
-	public String editPage(ProductDto product, RedirectAttributes rttr) {
+	public String editPage(ProductDto product, Principal principal, RedirectAttributes rttr) {
 		
-		boolean success = productService.updateProduct(product);
+		// 상품 정보 얻기
+		ProductDto oldProduct = productService.getProductById(product.getId());
 		
-		if (success) {
-			rttr.addFlashAttribute("message", "상품 정보가 수정되었습니다.");
+		// 상품 정보 작성자와 principal의 name과 비교해서 같을 때만 수정
+		if (oldProduct.getUser_id().equals(principal.getName())) {
+			
+			boolean success = productService.updateProduct(product);
+			
+			if (success) {
+				rttr.addFlashAttribute("message", "상품 정보가 수정되었습니다.");
+			} else {
+				rttr.addFlashAttribute("message", "상품 정보가 수정되지 않았습니다.");
+			}
+			
 		} else {
-			rttr.addFlashAttribute("message", "상품 정보가 수정되지 않았습니다.");
+			rttr.addFlashAttribute("message", "상품 정보 수정 권한이 없습니다.");
 		}
+		
+		rttr.addAttribute("id", product.getId());
 		
 		return "redirect:/product/detail";
 		
@@ -86,14 +102,27 @@ public class ProductController {
 	
 	// 상품 삭제
 	@PostMapping("remove")
-	public String remove(ProductDto product, RedirectAttributes rttr) {
+	public String remove(ProductDto product, Principal principal, RedirectAttributes rttr) {
 		
-		boolean success = productService.deleteProduct(product.getId());
+		// 상품 정보 얻기
+		ProductDto oldProduct = productService.getProductById(product.getId());
 		
-		if (success) {
-			rttr.addFlashAttribute("message", "상품 정보가 삭제되었습니다.");
+		// 상품 정보 작성자와 principal의 name과 비교해서 같을 때만 삭제
+		if (oldProduct.getUser_id().equals(principal.getName())) {
+			
+			boolean success = productService.deleteProduct(product.getId());
+			
+			if (success) {
+				rttr.addFlashAttribute("message", "상품 정보가 삭제되었습니다.");
+			} else {
+				rttr.addFlashAttribute("message", "상품 정보가 삭제되지 않았습니다.");
+			}
+			
 		} else {
-			rttr.addFlashAttribute("message", "상품 정보가 삭제되지 않았습니다.");
+			
+			rttr.addFlashAttribute("message", "상품 정보 삭제 권한이 없습니다.");
+			rttr.addAttribute("id", product.getId());
+			return "redirect:/product/detail";
 		}
 		
 		return "redirect:/product/search";
